+++ src/main/java/com/c4cometrue/mystorage/folder/FolderDataHandlerService.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import com.c4cometrue.mystorage.exception.ErrorCode;
