+++ src/main/java/com/c4cometrue/mystorage/folder/dto/FolderSummaryReq.java
@@ -1,0 +1,7 @@
+package com.c4cometrue.mystorage.folder.dto;
+
+public record FolderSummaryReq(
+        long folderId,
+        long userId
+) {
+}
