+++ src/test/java/com/c4cometrue/mystorage/folder/FolderControllerTest.java
@@ -3,6 +3,7 @@ package com.c4cometrue.mystorage.folder;
 import static com.c4cometrue.mystorage.TestConstants.*;
 import static org.mockito.Mockito.*;
 
+import com.c4cometrue.mystorage.folder.dto.FolderController;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
