+++ todo/Todo.txt
@@ -1,0 +1,42 @@
+# 구현 해야할 것
+- [] 일반 파일 요약
+- [] 루트 파일 요약
+
+# 요구 사항
+
+## 폴더 요약
+폴더 요약에는 다음 정보가 담겨야 한다
+1. 폴더 명
+2. 생성 날짜
+2. 수정 날짜 (하위 폴더 과업 발생 시에도 바뀌어야 한다)
+
+루트 폴더의 경우에는
+1. 하위 폴더 개수
+2. 하위 파일 개수
+3. 전체 파일 용량
+4. 가용 용량 (즉, 전체 파일의 크기가 2GB를 넘어간다면, 더 이상 파일 업로드가 불가능 합니다.)
+
+
+# 내 설계
+1. 파일 경로에 루트 폴더 이름(서버 저장명)이 들어간다
+2. 루트 폴더는 하나만 생성 가능하다
+3. 사용자 로그인 로직 만들면서 루트 폴더 생성하기
+
+## 작업 순위 및 설계
+1. 사용자 로그인 (루트 로직 생성 로직)
+2. 기존 폴더 생성 로직에서 parentId Long -> long
+
+3. 일반 폴더 요약 로직 만들기 (하위 폴더에서 변경 사항 생기면 상위 폴더 타서 수정 날짜 변경)
+
+4. 루트 폴더 요약 로직 만들기
+4-0. 테이블 분리 `루트 테이블 생성`, 폴더 테이블이 루트 테이블을 의존한다. 즉, 루트 테이블을 위한 별도의 서비스 레이어는 생성하지 않는 다
+4-1. 필드 : 유저 id(많이 조회되는 것이니 파일과 폴더에도 놔둔다), 경로, 최대 용량 (본 프로젝트에선 2GB로 제한, 별도의 enum , 상품 테이블 ㄴㄴ), 사용 용량
+루트 폴더 저장 경로(루트 폴더 저장 경로는 UUID + zoneTime -> 이 이유는 단순 UUID 만 할 경우 4-2 4-3 로직에서 파일과 폴더 개수를 체크하는 데 예상치 못한 문제가 발생 할 수 있기 때문 (다른 사용자의 폴더, 파일 개수를 체킹) 사용 용량도 이런 문제가 발생할 수 있기에)
+4-2. 하위 파일 개수 -> 서버 저장 경로 인덱스 설정 그리고 검색 (루트 폴더는 스토리지위치/루트폴더 로 되어 있고 나머지는 스토리지위치/루트폴더/파일서버저장명)
+4-3. 하위 폴더 개수 -> 서버 저장 경로 인덱스 설정 그리고 검색 (루트 폴더는 스토리지위치/루트폴더 로 되어 있고 나머지는 스토리지위치/루트폴더/파일서버저장명)
+
+
+## 번외 : 루트 테이블을 만든 이유
+모든 폴더에 폴더 용량 필드 만들면 이에 따른 수정(parent Id를 타고타고 다 수정해야하니) (업로드, 삭제 시 수정) 이러면 아찔해지죠
+        -> `루트 폴더`만 폴더 가용 용량 필드 가지고 있게 만들기 (업로드, 삭제 시 수정)
+        -> 그리고 루트 폴더는 자주 조회되니 && 루트 폴더만 가진 필드를 위해 다른 필드는 null 이 들어갈 필요가 없으니
